# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lee Pump App API swagger documentation",
        "description": "This is a sample API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/check-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "send otp",
                "description": "check otp",
                "operationId": "checkOtp",
                "parameters": [
                    {
                        "name": "mobile_no",
                        "in": "query",
                        "description": "mobile_no without +60",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "description": "otp",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/auth/request-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "send otp",
                "description": "request otp with number",
                "operationId": "requestOtp",
                "parameters": [
                    {
                        "name": "mobile_no",
                        "in": "query",
                        "description": "mobile_no without +60",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country_id",
                        "in": "query",
                        "description": "country_id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "cn/en",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/auth/request-user-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "send otp",
                "description": "request otp with username",
                "operationId": "requestUserOtp",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "username",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "cn/en",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "send otp",
                "description": "request user otp",
                "operationId": "resetPassword",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "username",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "passcode",
                        "in": "query",
                        "description": "otp",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}